+++ gpaw/setup_data.py
@@ -31,12 +31,6 @@ class SetupData:
         self.zero_reference = zero_reference
         self.generator_version = generator_version
 
-        # Default filename if this setup is written
-        if name is None or name == 'paw':
-            self.stdfilename = f'{symbol}.{self.setupname}'
-        else:
-            self.stdfilename = f'{symbol}.{name}.{self.setupname}'
-
         self.filename = None  # full path if this setup was loaded from file
         self.fingerprint = None  # hash value of file data if applicable
 
@@ -118,6 +112,16 @@ class SetupData:
 
         if readxml:
             self.read_xml(world=world)
+
+    @property
+    def stdfilename(self):
+        """Default filename if this setup is written."""
+        assert self.symbol is not None
+        assert self.setupname is not None
+        if self.name is None or self.name == 'paw':
+            return f'{self.symbol}.{self.setupname}'
+        else:
+            return f'{self.symbol}.{self.name}.{self.setupname}'
 
     def __repr__(self):
         return ('{0}({symbol!r}, {setupname!r}, name={name!r}, '
@@ -477,6 +481,8 @@ class PAWXMLParser(xml.sax.handler.ContentHandler):
         if name == 'atom':
             Z = float(attrs['Z'])
             setup.Z = Z
+            assert setup.symbol is None or setup.symbol == attrs['symbol']
+            setup.symbol = attrs['symbol']
             assert setup.Z == Z
             setup.Nc = float(attrs['core'])
             Nv = float(attrs['valence'])
@@ -491,6 +497,8 @@ class PAWXMLParser(xml.sax.handler.ContentHandler):
                     setup.orbital_free = True
                 else:
                     assert attrs['type'] == 'GGA'
+            assert setup.setupname is None or setup.setupname == setup.xcname
+            setup.setupname = setup.xcname
         elif name == 'ae_energy':
             setup.e_total = float(attrs['total'])
             setup.e_kinetic = float(attrs['kinetic'])